Database migrations: Add ability to publish migrations
Before this commit, it was not possible to publish the default
migrations for this tool using `php artisan vendor:publish --tag=`.

Once this commit has been applied, users will be able to export the
migrations using

```
php artisan vendor:publish --tag=telescope-migrations
```

This is similar to the behaviour seen in Laravel Passport.